+++ src/views/counter/Counter.vue
@@ -1,6 +1,13 @@
 <template>
-  <h2>Provide {{ count }}</h2>
-  <CounterAction></CounterAction>
+  <a-tabs :defaultActiveKey="activeKey" :animated="false">
+    <a-tab-pane key="PROVIDE_COUNT" tab="Provide Count">
+      <h2>PROVIDE_COUNT {{ count }}</h2>
+      <CounterAction></CounterAction>
+    </a-tab-pane>
+    <a-tab-pane key="VUEX_COUNT" tab="Vuex Count">
+      <h2>VUEX_COUNT {{ count }}</h2>
+    </a-tab-pane>
+  </a-tabs>
 </template>
 
 <script>
@@ -11,8 +18,10 @@ export default {
   components: {
     CounterAction,
   },
-  setup() {
+  setup(props) {
+    // count
     const count = ref(0);
+
     const add = () => {
       count.value++;
     };
@@ -27,8 +36,14 @@ export default {
     provide("addHandle", add);
     provide("minusHandle", minus);
 
+    // tab default key
+    const activeKey = ref("PROVIDE_COUNT");
+
+    console.log(props);
+
     return {
       count,
+      activeKey,
     };
   },
 };