+++ src/store/index.js
@@ -1,9 +1,11 @@
 import { createStore } from "vuex";
 import spaceX from "./modules/spaceX";
+import counter from "./modules/count";
 
 export default createStore({
   modules: {
     spaceX,
+    counter,
   },
 
   state: {