+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 --|--|--
  All | add git commit standard | [986a9f2](https://github.com/supermanbin/vueSamples/commit/986a9f2)
  all | provide & inject使用 | [81f4ae9](https://github.com/supermanbin/vueSamples/commit/81f4ae9)
+ counter tab | vue provide, inject & vuex count 使用方式 | [d66facc](https://github.com/supermanbin/vueSamples/commit/d66facc)
  global | optimize commit command | [f62a3aa](https://github.com/supermanbin/vueSamples/commit/f62a3aa)
  menu | 修改菜单 | [c4fb26d](https://github.com/supermanbin/vueSamples/commit/c4fb26d)
  revert 02 | revert 02 | [fec9904](https://github.com/supermanbin/vueSamples/commit/fec9904)
@@ -12,6 +13,7 @@
  revert | test git revert | [7c62b6b](https://github.com/supermanbin/vueSamples/commit/7c62b6b)
  revert | test revert | [a9bee5a](https://github.com/supermanbin/vueSamples/commit/a9bee5a)
  spaceX | spaceX | [f8ff03a](https://github.com/supermanbin/vueSamples/commit/f8ff03a)
+ yarn | add yarn | [e464bd2](https://github.com/supermanbin/vueSamples/commit/e464bd2)
 
 
 范围|描述|commitId